+++ packages/meditrak-server/src/database/models/DataSource.js
@@ -3,51 +3,8 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
 
-import {
-  DataSourceType as CommonDataSourceType,
-  DataSourceModel as CommonDataSourceModel,
-} from '@tupaia/database';
-
-export const DATA_SOURCE_SERVICE_TYPES = ['dhis', 'tupaia'];
-
-const getSurveyDateCode = surveyCode => `${surveyCode}SurveyDate`;
-
-export class DataSourceType extends CommonDataSourceType {
-  upsertSurveyDateElement = async () => {
-    this.assertFnCalledByDataGroup(this.upsertSurveyDateElement.name);
-
-    if (this.service_type !== this.SERVICE_TYPES.DHIS) {
-      // Non DHIS groups do not need a SurveyDate element
-      return;
-    }
-
-    const { id: dataElementId } = await this.model.updateOrCreate(
-      {
-        type: this.getTypes().DATA_ELEMENT,
-        code: getSurveyDateCode(this.code),
-      },
-      {
-        service_type: 'dhis',
-        config: { isDataRegional: this.config.isDataRegional },
-      },
-    );
-    await this.attachDataElement(dataElementId);
-  };
-
-  deleteSurveyDateElement = async () => {
-    this.assertFnCalledByDataGroup(this.deleteSurveyDateElement.name);
-
-    await this.model.delete({
-      type: this.getTypes().DATA_ELEMENT,
-      code: getSurveyDateCode(this.code),
-    });
-  };
-}
+import { DataSourceModel as CommonDataSourceModel } from '@tupaia/database';
 
 export class DataSourceModel extends CommonDataSourceModel {
   isDeletableViaApi = true;
-
-  get DatabaseTypeClass() {
-    return DataSourceType;
-  }
 }
